requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
networkx>=2.6.0
pyyaml>=5.4.0
asyncio>=3.4.3
logging>=0.4.9.6
scapy>=2.4.5
requests>=2.26.0
psutil>=5.9.0

# Machine Learning and Data Science
matplotlib>=3.5.0
seaborn>=0.11.0

# Web Dashboard
flask>=2.3.0
flask-socketio>=5.3.0
python-socketio>=5.8.0
eventlet>=0.33.0
psutil>=5.8.0
